Tighten action input typing and log selection
