+++ packages/shared/src/engine/optionalMarketEventEffects.ts
@@ -1,8 +1,4 @@
-import {
-  getTileByPosition,
-  getTilesBySector,
-  SECTOR_IDS,
-} from "../config/board.js";
+import { getTileByPosition } from "../config/board.js";
 import { startDeclineAuction } from "./auction.js";
 import type { AuctionResumePhase } from "./auctionTypes.js";
 import { hashSeed } from "./deckShuffle.js";
@@ -16,6 +12,10 @@ import {
   adjustCapital,
   transferCapital,
 } from "./marketEventPrimitives.js";
+import {
+  playerControllingMostTilesInAnySector,
+  playerWithFewestTiles,
+} from "./marketEventSelectors.js";
 import type { MarketEventTrigger } from "./marketEventTypes.js";
 import { getPlayer, transferTileOwnership } from "./stateUtils.js";
 
@@ -66,21 +66,6 @@ function ownDevelopmentTokensByPlayer(
   return totals;
 }
 
-function playerWithFewestTiles(
-  state: InternalGameState,
-): InternalPlayerState | null {
-  let best: InternalPlayerState | null = null;
-  for (const player of activePlayers(state)) {
-    if (
-      !best ||
-      player.ownedTilePositions.length < best.ownedTilePositions.length
-    ) {
-      best = player;
-    }
-  }
-  return best;
-}
-
 function mostExpensiveOwnedTile(
   state: InternalGameState,
   playerId: string,
@@ -101,28 +86,6 @@ function mostExpensiveOwnedTile(
     }
   }
   return bestPos;
-}
-
-function playerControllingMostTilesInAnySector(state: InternalGameState): {
-  playerId: string;
-  sectorId: string;
-  count: number;
-} | null {
-  let best: { playerId: string; sectorId: string; count: number } | null = null;
-  for (const player of activePlayers(state)) {
-    for (const sectorId of SECTOR_IDS) {
-      const sectorTiles = getTilesBySector(sectorId);
-      const owned = sectorTiles.filter((tile) =>
-        player.ownedTilePositions.some(
-          (pos) => String(pos) === String(tile.position),
-        ),
-      ).length;
-      if (!best || owned > best.count) {
-        best = { playerId: player.playerId, sectorId, count: owned };
-      }
-    }
-  }
-  return best && best.count > 0 ? best : null;
 }
 
 function pickSeededOwnedTile(
